+++ tcp_client/main/btd_http.c
@@ -1,0 +1,31 @@
+#include <string.h>
+#include "freertos/FreeRTOS.h" // FreeRTOS API
+#include "freertos/task.h"     // Task management
+#include "esp_wifi.h"
+#include "esp_log.h"
+#include "esp_event.h"
+
+
+static const char *TAG = "BTD_HTTP";
+
+esp_err_t start_wifi_ap(const char *ssid, const char *password)
+{
+    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
+    ESP_ERROR_CHECK(esp_wifi_init(&cfg));
+    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_AP));
+
+    wifi_config_t ap_config = {0};
+    strncpy((char *)ap_config.ap.ssid, ssid, sizeof(ap_config.ap.ssid) - 1);
+    strncpy((char *)ap_config.ap.password, password, sizeof(ap_config.ap.password) - 1);
+    ap_config.ap.ssid_len = strlen(ssid);
+    ap_config.ap.channel = 1;
+    ap_config.ap.authmode = WIFI_AUTH_WPA2_PSK;
+    ap_config.ap.max_connection = 4;
+
+    ESP_ERROR_CHECK(esp_wifi_set_config(WIFI_IF_AP, &ap_config));
+    return esp_wifi_start();
+}
+
+
+
+
