+++ tcp_client/main/btd_http.h
@@ -1,0 +1,10 @@
+
+#ifndef BT_HTTP_H
+#define BT_HTTP_H
+
+#include "esp_err.h"
+
+esp_err_t start_wifi_ap(const char *ssid, const char *password);
+
+#endif // BT_HTTP_H
+
