+++ tcp_client/main/btd_wifi.c
@@ -18,29 +18,14 @@ static const char *TAG = "BTD_WIFI";
 #define NVS_KEY_FPCOUNT "fp_count"
 #define NVS_FP_PREFIX "fp_"
 
-esp_err_t start_wifi_ap(const char *ssid, const char *password)
-{
-    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
-    ESP_ERROR_CHECK(esp_wifi_init(&cfg));
-    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_AP));
 
-    wifi_config_t ap_config = {0};
-    strncpy((char *)ap_config.ap.ssid, ssid, sizeof(ap_config.ap.ssid) - 1);
-    strncpy((char *)ap_config.ap.password, password, sizeof(ap_config.ap.password) - 1);
-    ap_config.ap.ssid_len = strlen(ssid);
-    ap_config.ap.channel = 1;
-    ap_config.ap.authmode = WIFI_AUTH_WPA2_PSK;
-    ap_config.ap.max_connection = 4;
 
-    ESP_ERROR_CHECK(esp_wifi_set_config(WIFI_IF_AP, &ap_config));
-    return esp_wifi_start();
-}
-
-esp_err_t stop_wifi_ap()
+esp_err_t stop_wifi()
 {
     ESP_ERROR_CHECK(esp_wifi_stop());
     return esp_wifi_deinit();
 }
+
 
 static void make_fingerprint_key(char *key, size_t key_size, uint8_t index)
 {
@@ -144,7 +129,7 @@ static esp_err_t scan_and_create_fingerprint(wifi_location_fingerprint_t *fp)
     add_location_name(fp);
     ESP_LOGI(TAG, "Location fingerprint created: %s", fp->name);
 
-    ESP_ERROR_CHECK(stop_wifi_ap());
+    ESP_ERROR_CHECK(stop_wifi());
     return ESP_OK;
 }
